pyproject.toml
[tool.poetry]
name = "mace-api"
version = "0.1.0"
description = "API to manage data for the MACE lab"
authors = ["Evan Thomas <evan.thomas@epfl.ch>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.104.1"
uvicorn = "^0.23.2"
sqlmodel = "^0.0.11"
geoalchemy2 = "^0.14.2"
aiosqlite = "^0.19.0"
httpx = "^0.25.1"
pytest-asyncio = "^0.21.1"
asyncpg = "^0.29.0"
shapely = "^2.0.2"
sqlalchemy-utils = "^0.41.1"
alembic = "^1.12.1"
pydantic-geojson = "^0.1.1"
setuptools = "^68.2.2"
wheel = "^0.41.3"
numpy = "^1.26.1"
geos = "^0.2.3"
gpxpy = "^1.6.1"
orjson = "^3.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-watcher = "^0.3.4"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
